updated jenkinsfile to include namespaces

+++ Jenkinsfile
@@ -92,7 +92,7 @@ spec:
         ephemeral-storage: "2Gi"
 
   - name: maven
-    image: 779921734503.dkr.ecr.eu-west-1.amazonaws.com/jnlp-dood-new-infra:200608
+    image: 779921734503.dkr.ecr.eu-west-1.amazonaws.com/jnlp-dood-new-infra:200710
     imagePullPolicy: IfNotPresent
     command: ['docker', 'run', '-p', '80:80', 'httpd:latest']
     tty: true
@@ -217,18 +217,28 @@ spec:
             dir ('Palisade-examples') {
                 git branch: 'develop', url: 'https://github.com/gchq/Palisade-examples.git'
                 git branch: GIT_BRANCH_NAME, url: 'https://github.com/gchq/Palisade-examples.git'
-                container('docker-cmds') {
+                container('maven') {
                     configFileProvider([configFile(fileId: "${env.CONFIG_FILE}", variable: 'MAVEN_SETTINGS')]) {
                         sh 'mvn -s $MAVEN_SETTINGS install -P quick'
-                        sh '''
-                            helm version
-                            bash deployment/local-k8s/local-bash-scripts/deployServicesToK8s.sh
-                            helm list
-                            kubectl get pods
-                            bash deployment/local-k8s/local-bash-scripts/runFormattedK8sExample.sh
-                            helm uninstall palisade
-                            bash deployment/local-k8s/local-bash-scripts/verify.sh
-                        '''
+                        if (sh(script: "namespace-create test", returnStatus: true) == 0) {
+                            if (sh(script: "helm upgrade --install palisade . " +
+                                    "--set global.persistence.dataStores.palisade-data-store.local.hostPath=$(pwd)/resources/data  " +
+                                    "--set global.persistence.classpathJars.local.hostPath=$(pwd)/deployment/target " +
+                                    "--namespace test", returnStatus: true) == 0) {
+                                echo("successfully deployed")
+                                sh '''
+                                    helm version
+                                    helm list
+                                    kubectl get pods
+                                    bash deployment/local-k8s/local-bash-scripts/runFormattedK8sExample.sh
+                                    helm uninstall palisade
+                                    bash deployment/local-k8s/local-bash-scripts/verify.sh
+                                '''
+                            } else {
+                                error("Build failed because of failed maven deploy")
+                            }
+
+                        }
                     }
                 }
             }
